Adds system benchmarks command (#1232)
* Refactor benchmark command to make it easier to extend

* Add new command and read config

* Set up service and generator

* Improve policy creation and add docker service deployer

* Bump es api version

* Add new options

* Add new reportable types

* Collect and report summary metrics

* Add documentation

* Add system benchmark test

* Save report file

* Ensure all data is generated previous to agents being assigned

* Save file indented

* Update readme

* Lint

* Make test faster

* Go back to 7.17 client

* fix tests

* Change comments

* Add monitoringenabled option

* Update report.go

* Delete outputs.go from kibana

* Rename reports payload property

* Remove pkg/errors from ingest, kibana, and servicedeployer

* Amend const and comment

* Avoid overriding configs unadvertedly

* Change byte replace to strings.Replacer

* Undo change in report benchmark command

* Use policy template name to generate input name

* Fix readme

* Add corpora file to report

* Cleanup corpora generated data

* Fix lint error

* Fix test package format

* Fix lint errors

* Fix processor error return

* Rename folder to accomodate to spec

* Add prefix check

+++ scripts/test-check-packages.sh
@@ -69,18 +69,23 @@ for d in test/packages/${PACKAGE_TEST_TYPE:-other}/${PACKAGE_UNDER_TEST:-*}/; do
     elastic-package install -v
 
     if [ "${PACKAGE_TEST_TYPE:-other}" == "benchmarks" ]; then
-      rm -rf "${OLDPWD}/build/benchmark-results"
-      elastic-package benchmark -v --report-format xUnit --report-output file --fail-on-missing
-      
-      rm -rf "${OLDPWD}/build/benchmark-results-old"
-      mv "${OLDPWD}/build/benchmark-results" "${OLDPWD}/build/benchmark-results-old"
-      
-      elastic-package benchmark -v --report-format json --report-output file --fail-on-missing
-      
-      elastic-package report --fail-on-missing benchmark \
-        --new ${OLDPWD}/build/benchmark-results \
-        --old ${OLDPWD}/build/benchmark-results-old \
-        --threshold 1 --report-output-path="${OLDPWD}/build/benchreport"
+      if [ "${PACKAGE_UNDER_TEST:-*}" == "pipeline_benchmark" ]; then
+        rm -rf "${OLDPWD}/build/benchmark-results"
+        elastic-package benchmark pipeline -v --report-format xUnit --report-output file --fail-on-missing
+        
+        rm -rf "${OLDPWD}/build/benchmark-results-old"
+        mv "${OLDPWD}/build/benchmark-results" "${OLDPWD}/build/benchmark-results-old"
+        
+        elastic-package benchmark pipeline -v --report-format json --report-output file --fail-on-missing
+        
+        elastic-package report --fail-on-missing benchmark \
+          --new ${OLDPWD}/build/benchmark-results \
+          --old ${OLDPWD}/build/benchmark-results-old \
+          --threshold 1 --report-output-path="${OLDPWD}/build/benchreport"
+      fi
+      if [ "${PACKAGE_UNDER_TEST:-*}" == "system_benchmark" ]; then
+        elastic-package benchmark system --benchmark logs-benchmark -v --defer-cleanup 1s
+      fi
     else
       # defer-cleanup is set to a short period to verify that the option is available
       elastic-package test -v --report-format xUnit --report-output file --defer-cleanup 1s --test-coverage
